Move AbstractOffsetNode to the correct namespace
