+++ SAGESharp/IO/TreeBasedSerialization.cs
@@ -494,40 +494,5 @@ namespace SAGESharp.IO
 
         private static bool IsType(object value) => typeof(T) == value.GetType();
     }
-
-    internal abstract class AbstractOffsetNode : IOffsetNode
-    {
-        protected AbstractOffsetNode(IDataNode childNode)
-        {
-            Validate.ArgumentNotNull(childNode, nameof(childNode));
-
-            ChildNode = childNode;
-        }
-
-        public IDataNode ChildNode { get; }
-
-        public uint ReadOffset(IBinaryReader binaryReader)
-        {
-            Validate.ArgumentNotNull(binaryReader, nameof(binaryReader));
-
-            return binaryReader.ReadUInt32();
-        }
-
-        public abstract uint Write(IBinaryWriter binaryWriter, object value);
-
-        protected uint WriteOffset(IBinaryWriter binaryWriter)
-        {
-            if (binaryWriter.Position > uint.MaxValue)
-            {
-                throw new InvalidOperationException("Offset is bigger than 4 bytes.");
-            }
-
-            uint offsetPosition = (uint) binaryWriter.Position;
-
-            binaryWriter.WriteUInt32(0);
-
-            return offsetPosition;
-        }
-    }
     #endregion
 }
